Adjust log calls to use logf insead

+++ src/OTA_Handler.cpp
@@ -80,7 +80,7 @@ void OTA_Handler::Start_Firmware_Update(const OTA_Update_Callback& fw_callback, 
     m_fw_updater = m_fw_callback->Get_Updater();
 
     if (!m_publish_callback || !m_send_fw_state_callback || !m_finish_callback || !m_fw_updater) {
-      m_logger.log(OTA_CB_IS_NULL);
+      m_logger.logf(OTA_CB_IS_NULL);
       (void)m_send_fw_state_callback(FW_STATE_FAILED, OTA_CB_IS_NULL);
         return Handle_Failure(OTA_Failure_Response::RETRY_NOTHING);
     }
@@ -90,7 +90,7 @@ void OTA_Handler::Start_Firmware_Update(const OTA_Update_Callback& fw_callback, 
 void OTA_Handler::Stop_Firmware_Update() {
     m_watchdog.detach();
     m_fw_updater->reset();
-    m_logger.log(FW_UPDATE_ABORTED);
+    m_logger.logf(FW_UPDATE_ABORTED);
     (void)m_send_fw_state_callback(FW_STATE_FAILED, FW_UPDATE_ABORTED);
     Handle_Failure(OTA_Failure_Response::RETRY_NOTHING);
     m_fw_callback = nullptr;
@@ -112,7 +112,7 @@ void OTA_Handler::Process_Firmware_Packet(const size_t& current_chunk, uint8_t *
     if (current_chunk == 0U) {
         // Initialize Flash
         if (!m_fw_updater->begin(m_fw_size)) {
-          m_logger.log(ERROR_UPDATE_BEGIN);
+          m_logger.logf(ERROR_UPDATE_BEGIN);
           (void)m_send_fw_state_callback(FW_STATE_FAILED, ERROR_UPDATE_BEGIN);
           return Handle_Failure(OTA_Failure_Response::RETRY_UPDATE);
         }
@@ -123,7 +123,7 @@ void OTA_Handler::Process_Firmware_Packet(const size_t& current_chunk, uint8_t *
     if (written_bytes != total_bytes) {
         char message[Helper::detectSize(ERROR_UPDATE_WRITE, written_bytes, total_bytes)];
         snprintf(message, sizeof(message), ERROR_UPDATE_WRITE, written_bytes, total_bytes);
-        m_logger.log(message);
+        m_logger.logf(message);
         (void)m_send_fw_state_callback(FW_STATE_FAILED, message);
         return Handle_Failure(OTA_Failure_Response::RETRY_UPDATE);
     }
@@ -164,7 +164,7 @@ void OTA_Handler::Request_Next_Firmware_Packet() {
     }
 
     if (!m_publish_callback(m_requested_chunks)) {
-      m_logger.log(UNABLE_TO_REQUEST_CHUNCKS);
+      m_logger.logf(UNABLE_TO_REQUEST_CHUNCKS);
       (void)m_send_fw_state_callback(FW_STATE_FAILED, UNABLE_TO_REQUEST_CHUNCKS);
     }
 
@@ -187,23 +187,23 @@ void OTA_Handler::Finish_Firmware_Update() {
     // Check if the initally received checksum is the same as the one we calculated from the received binary data,
     // if not we assume the binary data has been changed or not completly downloaded --> Firmware update failed.
     if (memcmp(m_fw_checksum, calculated_hash, sizeof(m_fw_checksum)) == 0) {
-        m_logger.log(CHKS_VER_FAILED);
+        m_logger.logf(CHKS_VER_FAILED);
         (void)m_send_fw_state_callback(FW_STATE_FAILED, CHKS_VER_FAILED);
         return Handle_Failure(OTA_Failure_Response::RETRY_UPDATE);
     }
 
 #if THINGSBOARD_ENABLE_DEBUG
-    m_logger.log(CHKS_VER_SUCCESS);
+    m_logger.logf(CHKS_VER_SUCCESS);
 #endif // THINGSBOARD_ENABLE_DEBUG
 
     if (!m_fw_updater->end()) {
-        m_logger.log(ERROR_UPDATE_END);
+        m_logger.logf(ERROR_UPDATE_END);
         (void)m_send_fw_state_callback(FW_STATE_FAILED, ERROR_UPDATE_END);
         return Handle_Failure(OTA_Failure_Response::RETRY_UPDATE);
     }
 
 #if THINGSBOARD_ENABLE_DEBUG
-    m_logger.log(FW_UPDATE_SUCCESS);
+    m_logger.logf(FW_UPDATE_SUCCESS);
 #endif // THINGSBOARD_ENABLE_DEBUG
     (void)m_send_fw_state_callback(FW_STATE_UPDATING, nullptr);
